removed squarex concept, now its inferred based on the x. sured up support for the client server moving at the same time. game server does NOT yet send the path properly for newly joined neighbors

+++ Pather.Client/GameFramework/FrontEndTickManager.cs
@@ -82,7 +82,7 @@ namespace Pather.Client.GameFramework
             }
         }
 
-        public override void SetServerLatency(long latency)
+        public override void SetServerLatency(int latency)
         {
             base.SetServerLatency(latency);
             hasLatency = true;
